fixes and improvements
chore: add assets folder (images, styles) to shared

chore: remove template stories directory

chore: remove empty sass options from next.config.ts

chore: rename share slice to shared

refactor: move Button component and styles into ui directory

docs: add JSDoc annotations to Button component and props

refactor: move variables styles file into shared/assets/styles

docs(storybook): update Button stories
