+++ .storybook/preview.ts
@@ -9,6 +9,13 @@ const preview: Preview = {
         date: /Date$/i,
       },
     },
+    backgrounds: {
+      options: {
+        light: { name: "Light", value: "#f8f8f8" },
+        dark: { name: "Dark", value: "#333333" },
+        cyan: { name: "Cyan", value: "#0FCCD3" },
+      },
+    },
   },
 }
