+++ next.config.ts
@@ -2,9 +2,6 @@ import type { NextConfig } from "next"
 
 const nextConfig: NextConfig = {
   output: "standalone",
-  sassOptions: {
-    // additionalData: `@use "styles/variables" as *;`,
-  },
 }
 
 export default nextConfig
