+++ src/shared/assets/styles/_variables.module.scss
@@ -6,7 +6,7 @@ $color-primary-blue: #0FCCD3;
 $color-primary-white: #E8F8F8;
 /* "основной цвет для заголовков" */
 $color-text-blue: #0EB3B3;
-/* "основной цвет для обычно текста" */
+/* "основной цвет для обычного текста" */
 $color-text-dark: #0E3F53;
 
 :root {
