+++ src/shared/ui/Button/story/Button.stories.module.scss
@@ -1,0 +1,8 @@
+.storyDecorator {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  padding: 0 50px;
+  height: 100vh;
+}
