+++ src/shared/ui/Button/ui/Button.module.scss
@@ -1,5 +1,5 @@
 .root {
-  padding: 5px 30px;
+  padding: 5px 25px;
   min-height: 36px;
 
   display: flex;
@@ -37,11 +37,6 @@
 }
 
 .glass {
-  border-color: transparent;
-
-  /* кнопки */
-  box-shadow: -2px -2px 4px rgba(255, 255, 255, 0.25), 1px 2px 3.9px rgba(9, 25, 20, 0.28), inset 2px 1px 4px rgba(14, 47, 32, 0.23), inset -2px -3px 4px rgba(255, 255, 255, 0.5);
-
   &:hover {
     background-color: var(--color-primary-white);
     color: var(--color-text-dark);
